// src/api/generated/models/assessmentStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * ECD Backend API
 * Early Childhood Development management system API.
 * OpenAPI spec version: 1.0
 */

export type AssessmentStatus = typeof AssessmentStatus[keyof typeof AssessmentStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AssessmentStatus = {
  draft: 'draft',
  submitted: 'submitted',
  verified: 'verified',
  rejected: 'rejected',
} as const;